+++ tests/test_parser.py
@@ -3,46 +3,7 @@ import os
 import pytest
 from nomad.client import normalize_all, parse
 from nomad.units import ureg
-
-
-def set_monkey_patch(monkeypatch):
-    def mockreturn_search(*args, upload_id=None):
-        return None
-
-    monkeypatch.setattr(
-        'nomad_hysprint.parsers.hysprint_measurement_parser.set_sample_reference',
-        mockreturn_search,
-    )
-
-    monkeypatch.setattr(
-        'nomad_hysprint.schema_packages.hysprint_package.set_sample_reference',
-        mockreturn_search,
-    )
-
-
-def delete_json():
-    for file in os.listdir(os.path.join('tests/data')):
-        if not file.endswith('archive.json'):
-            continue
-        os.remove(os.path.join('tests', 'data', file))
-
-
-delete_json()
-
-
-def get_archive(file_base, monkeypatch):
-    set_monkey_patch(monkeypatch)
-    file_name = os.path.join('tests', 'data', file_base)
-    file_archive = parse(file_name)[0]
-    assert file_archive.data
-
-    for file in os.listdir(os.path.join('tests/data')):
-        if 'archive.json' not in file:
-            continue
-        measurement = os.path.join('tests', 'data', file)
-        measurement_archive = parse(measurement)[0]
-
-    return measurement_archive
+from utils import delete_json, get_archive
 
 
 @pytest.fixture(
